made base class for org tests
