+++ startsmart/organizations/tests/test_views.py
@@ -1,83 +1,10 @@
-import random
-
 from django.urls import reverse
-from django.test import TestCase
-from django.contrib.auth.models import User
 
 from ..models import Organization, Post
-
-class TestModels(TestCase):
-    def setUp(self):
-        '''
-        Must follow this order
-        '''
-        self.owner = self._create_organization_owner()
-        self.organization = self._create_organization()
-        self.owner_posts = self._create_posts_from_owner()
-
-    def tearDown(self):
-        '''
-        Delete all objects created
-        '''
-
-        self.owner.delete()
-
-    def _create_organization_owner(self):
-        '''
-        Create the User that will own the organization.
-        Save the attributes so we can test that the user exists
-        '''
-        self.username = "testing"
-        self.email = "test@email.com"
-        self.password = "testing#2020"
-
-        owner = User.objects.create_user(
-            username = self.username,
-            password = self.password,
-            email=self.email
-        )
-
-        return owner
+from .utils import BaseTestOrganizations
 
 
-
-    def _create_posts_from_owner(self):
-        '''
-        Create posts within the test organization authored by the owner
-        '''
-        self.post_texts = [
-            "Hello world",
-            "This is a test post",
-        ]
-
-        posts = []
-
-        for text in self.post_texts:
-            post = Post.objects.create(
-                author = self.owner,
-                organization = self.organization,
-                text = text,
-            )
-            posts.append(posts)
-
-        return posts
-
-    def _create_organization(self):
-        '''
-        Create an organization.
-        Save the attributes to test that the organization exists
-        '''
-        self.organization_name = "Test Org Inc."
-        self.organization_description = "for testing"
-
-        organization = Organization.objects.create(
-            owner = self.owner,
-            name = self.organization_name,
-            description = self.organization_description,
-        )
-
-
-        return organization
+class TestViews(BaseTestOrganizations):
 
     def test_view_organization(self):
         '''
@@ -102,6 +29,20 @@ class TestModels(TestCase):
         '''
         self.client.login(username = self.username, password=self.password)
 
+        response = self.client.get(reverse(
+            "delete-organization",
+            kwargs = {
+                "organization_name": self.organization_name
+            }
+        ))
+
+        # verify that the organization to be deleted is the organization
+        # we just made
+        organization = response.context['organization']
+        assert self.organization_name == organization.name
+
+
+
         response = self.client.post(reverse(
             "delete-organization",
             kwargs = {
@@ -120,6 +61,8 @@ class TestModels(TestCase):
             name = self.organization_name,
         )
 
+        # make org again
+        self.organization = self._create_organization()
         self.client.logout()
